Rewrite data.bitvec to use new representation.
Data.bitvec from the standard library uses an inefficient internal
representation of vector of bools. This change rewrites the internals
(with the intention of keeping the existing API) so that bitvec is now a
subtype of nat.

This change also includes some minor refactors to the project structure:

  * Move dec_trivial_tac to a tactic module
  * Add a nat directory with a lemmas module, for project specific nat
  lemmas.

More operations will need to be added for reopt, but this should get us
to rough feature parity with data.bitvec from the standard library.