feat: init rel spec for SMT terms and their model denot
This commit adds an `Interp` relation which, for a given typing
context and model, relates an SMT term to it's interpretation
in that model.

One remaining issue (I believe) is that the interpretation of
`forall` and `exists` quantifies over all possible denotational
values for the variable in question, which combined with (1)
our current Lean4 denotation of SMT builtin functions
`eq` and `ite` relies on definitional equality
of our denotational values (which could be nice later for proofs)
and (2) our denotation of arrays admits observing the non-structural
equality of extensionally equal arrays. Possible solutions are
to (a) not use definitional equality or (b) modify the denotation
of arrays to be a subset of the current denotation which enforces
the implied well-formedness constraints (i.e., so extensional
and definitional equality in fact coincide).