WIP proofs for Command denotation
Initial approach for Command denotation is that it updates
the current SMT context (i.e., the Env, in scope asserts,
and function definitions). That lead to a bunch of proofs
regarding the well-formedness of Terms when extending
the environment. They're not too bad, but they're a little
tedious writing manually as Lean4 functions. Anyway,
I left off near the tail end of `Term.WS.updMap (i.e., that
if a new variable is added to the environment any terms
that were well formed are still well-formed in that new
environment). If that finishes nicely, I've got just a few
lemmas to prove (currrently axioms) regarding environment extension
in some particular cases.